+++ lib/src/platform/cookie_reader.dart
@@ -1,0 +1,2 @@
+export 'cookie_reader_stub.dart'
+    if (dart.library.js_interop) 'cookie_reader_web.dart';